+++ src/HttpKernel/Controller/ControllerResolver.php
@@ -5,7 +5,7 @@
  * Copyright (c) 2015 Tomas Votruba (http://tomasvotruba.cz).
  */
 
-namespace Zenify\ControllerAutowire\HttpKernel;
+namespace Zenify\ControllerAutowire\HttpKernel\Controller;
 
 use Symfony\Component\DependencyInjection\ContainerAwareInterface;
 use Symfony\Component\DependencyInjection\ContainerInterface;
