+++ tests/CompleteTest.php
@@ -7,11 +7,11 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\DependencyInjection\Exception\ServiceNotFoundException;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Tests\Controller;
-use Zenify\ControllerAutowire\HttpKernel\ControllerResolver;
+use Zenify\ControllerAutowire\HttpKernel\Controller\ControllerResolver;
 use Zenify\ControllerAutowire\Tests\CompleteTestSource\DoNotScan\SomeRegisteredController;
 use Zenify\ControllerAutowire\Tests\CompleteTestSource\Scan\ContainerAwareController;
-use Zenify\ControllerAutowire\Tests\HttpKernel\ControllerFinderSource\SomeController;
-use Zenify\ControllerAutowire\Tests\HttpKernel\ControllerFinderSource\SomeService;
+use Zenify\ControllerAutowire\Tests\HttpKernel\Controller\ControllerFinderSource\SomeController;
+use Zenify\ControllerAutowire\Tests\HttpKernel\Controller\ControllerFinderSource\SomeService;
 
 final class CompleteTest extends PHPUnit_Framework_TestCase
 {
