+++ tests/HttpKernel/Controller/ControllerFinderSource/SomeControllerFactory.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Zenify\ControllerAutowire\Tests\HttpKernel\Controller\ControllerFinderSource;
+
+class SomeControllerFactory
+{
+}
