+++ tests/HttpKernel/Controller/ControllerFinderSource/SomeOtherController.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Zenify\ControllerAutowire\Tests\HttpKernel\Controller\ControllerFinderSource;
+
+class SomeOtherController
+{
+}
