+++ tests/HttpKernel/Controller/ControllerFinderSource/SomeService.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Zenify\ControllerAutowire\Tests\HttpKernel\Controller\ControllerFinderSource;
+
+class SomeService
+{
+}
