+++ tests/HttpKernel/Controller/ControllerFinderSource/Tests/SomeControllerTest.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Zenify\ControllerAutowire\Tests\HttpKernel\Controller\ControllerFinderSource\Tests;
+
+class SomeControllerTest
+{
+}
