--- tests/HttpKernel/ControllerFinderSource/SomeClass.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Zenify\ControllerAutowire\Tests\HttpKernel\ControllerFinderSource;
-
-class SomeClass
-{
-}
