--- tests/HttpKernel/ControllerFinderSource/SomeControllerFactory.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Zenify\ControllerAutowire\Tests\HttpKernel\ControllerFinderSource;
-
-class SomeControllerFactory
-{
-}
