--- tests/HttpKernel/ControllerFinderSource/SomeService.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Zenify\ControllerAutowire\Tests\HttpKernel\ControllerFinderSource;
-
-class SomeService
-{
-}
