--- tests/HttpKernel/ControllerFinderSource/Tests/SomeControllerTest.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Zenify\ControllerAutowire\Tests\HttpKernel\ControllerFinderSource\Tests;
-
-class SomeControllerTest
-{
-}
