+++ tests/Resources/config/config.yml
@@ -4,12 +4,12 @@ parameters:
 
 services:
     some.service:
-        class: Zenify\ControllerAutowire\Tests\HttpKernel\ControllerFinderSource\SomeService
+        class: Zenify\ControllerAutowire\Tests\HttpKernel\Controller\ControllerFinderSource\SomeService
     some.controller.service:
         class: Zenify\ControllerAutowire\Tests\CompleteTestSource\DoNotScan\SomeRegisteredController
 
 
 zenify_controller_autowire:
     controller_dirs:
-        - %kernel.root_dir%/HttpKernel/ControllerFinderSource
+        - %kernel.root_dir%/HttpKernel/Controller/ControllerFinderSource
         - %kernel.root_dir%/CompleteTestSource/Scan
